test: add UseConnection source generator unit tests
